refactor(sct_events.py): Crating object for group common DB events
* Trello - https://trello.com/c/V3Ca21uD
   Trello description:
    Same filtered events are used for common actions.
    Need to find a way to group them in one file under the same Object
    Example:
    ```
      with DbEventsFilter(type='DATABASE_ERROR', line="repair's stream failed: streaming::stream_exception",
                     node=self.target_node), \
         DbEventsFilter(type='RUNTIME_ERROR', line='Can not find stream_manager', node=self.target_node), \
         DbEventsFilter(type='RUNTIME_ERROR', line='is aborted', node=self.target_node), \
         DbEventsFilter(type='RUNTIME_ERROR', line='Failed to repair', node=self.target_node):
    ```
